requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
Deprecated==1.2.14
Django==3.2.19
django-celery-beat==2.2.0
django-celery-results==2.1.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-redis==5.2.0
django-rest-swagger==2.2.0
django-timezone-field==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.3.0
MarkupSafe==2.1.3
mysqlclient==2.1.1
openapi-codec==1.3.2
packaging==23.1
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.38
PyJWT==2.7.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.1.4
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
zipp==3.15.0
